fixed incorrect function names

+++ common/point.c
@@ -25,14 +25,14 @@ Point4i point4i(int w, int x, int y, int z)
 	return p;
 }
 
-Point2f point2f(int x, int y)
+Point2f point2f(float x, float y)
 {
 	Point2f p;
 	p.x = x;
 	p.y = y;
 	return p;
 }
-Point3f point3f(int x, int y, int z)
+Point3f point3f(float x, float y, float z)
 {
 	Point3f p;
 	p.x = x;
@@ -40,7 +40,7 @@ Point3f point3f(int x, int y, int z)
 	p.z = z;
 	return p;
 }
-Point4f point4f(int w, int x, int y, int z)
+Point4f point4f(float w, float x, float y, float z)
 {
 	Point4f p;
 	p.w = w;
